organizing code and retrieving carb ratios
