+++ FreeAPS/Sources/Modules/Bolus/BolusProvider.swift
@@ -4,6 +4,12 @@ extension Bolus {
             storage.retrieve(OpenAPS.Enact.suggested, as: Suggestion.self)
         }
 
+        func getProfile() -> CarbRatios {
+            storage.retrieve(OpenAPS.Settings.carbRatios, as: CarbRatios.self)
+                ?? CarbRatios(from: OpenAPS.defaults(for: OpenAPS.Settings.carbRatios))
+                ?? CarbRatios(units: .grams, schedule: [])
+        }
+
         func pumpSettings() -> PumpSettings {
             storage.retrieve(OpenAPS.Settings.settings, as: PumpSettings.self)
                 ?? PumpSettings(from: OpenAPS.defaults(for: OpenAPS.Settings.settings))
